tidy up the test helper
extract test model definitions into separate files to be required as
necessary
